backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
PyPDF2==3.0.1
python-docx==1.2.0
pdfplumber==0.11.8
groq==1.0.0
pydantic==2.12.5
python-dotenv==1.2.1
google-generativeai==0.8.3